feat: implementar createBlock(), utilitário stringToArray(), arquivo de erros e atualização para Uint32
BlogFsCore:
- Criado método createBlock(label: string) responsável por registrar novos blocos
  na região de metadados. A lógica inclui:
    - verificação do limite MAX_BLOCKS
    - cálculo do endereço livre do bloco
    - escrita do label em bytes (15 bytes UTF-8 / preenchimento com 0x00)
    - atualização do contador total de blocos
    - definição do status do bloco
    - cálculo e gravação do endereço inicial da sessão de registros para o bloco

- Criado método privado stringToArray(string) → Uint8[] para conversão de labels
  em bytes UTF-8, padronizando a escrita em meta_view.

- Adicionado método privado save_metadata_data(), responsável por persistir no
  arquivo físico todas as alterações feitas nos metadados.

Interfaces.ts / Limits.ts:
- Atualizado o tipo de r_session_address de BigUint64Array para Uint32Array,
  eliminando o uso desnecessário de endereços 64-bit e alinhando o sistema ao
  limite prático de arquivos até 4GB.

Erros.ts:
- Arquivo criado para centralizar erros customizados, como LimitedBlockReached,
  mantendo padronização e clareza dentro da classe.

+++ src/Interfaces.ts
@@ -9,7 +9,7 @@ export interface meta_block_session {
    label: string;     // (15 bytes UTF-8): identificação do bloco
    stats: Uint8Array; // status (soft delete)
    r_count: Uint32Array;             // quantidade de registros
-   r_session_address:BigUint64Array; // endereço inicial da sessão de registros"
+   r_session_address:Uint32Array; // endereço inicial da sessão de registros"
 }
 
 export interface meta_register_session {
